Added DECLSPEC
Classes required a USML_DECLSPEC decleration in the header for windows
systems.

+++ types/data_grid.h
@@ -230,13 +230,13 @@ class USML_DLLEXPORT data_grid {
             }
 
             NcVar* earth_radius = _file->add_var( "earth_radius", ncDouble, 0 ) ;
-            for(long i=0; i < NUM_DIMS; ++i) {
+            for(size_t i=0; i < NUM_DIMS; ++i) {
                 std::stringstream ss ;
                 ss << "axis" << i << "\0" ;
-                axis_dim[i] = _file->add_dim( (ss.str()).c_str(), _axis[i]->size() ) ;
+                axis_dim[i] = _file->add_dim( (ss.str()).c_str(), (long)_axis[i]->size() ) ;
                 axis_size[i] = axis_dim[i] ;
                 axis_var[i] = _file->add_var( (ss.str()).c_str(), type, axis_dim[i] ) ;
-                data_size[i] = _axis[i]->size() ;
+                data_size[i] = (long)_axis[i]->size() ;
             }
             NcVar* data_var = _file->add_var( "data", type, NUM_DIMS, &*axis_size ) ;
